fully completed the basis of the controller and business layer tests. some tests still need certain debugging.

+++ src/main/java/com/rooshdashboard/rooshdashboard/business/impl/account/DeleteAccountUseCaseImpl.java
@@ -1,6 +1,7 @@
 package com.rooshdashboard.rooshdashboard.business.impl.account;
 
 import com.rooshdashboard.rooshdashboard.business.IAccount.DeleteAccountUseCase;
+import com.rooshdashboard.rooshdashboard.business.exception.InvalidAccountException;
 import com.rooshdashboard.rooshdashboard.domain.Account.DeleteAccountResponse;
 import com.rooshdashboard.rooshdashboard.persistance.AccountRepository;
 import lombok.AllArgsConstructor;
@@ -13,6 +14,9 @@ public class DeleteAccountUseCaseImpl implements DeleteAccountUseCase {
 
     @Override
     public DeleteAccountResponse deleteAccount(long accountId) {
+        if(!accountRepository.existsById(accountId)){
+            throw new InvalidAccountException("ACCOUNT_NOT_FOUND");
+        }
         this.accountRepository.deleteById(accountId);
 
         return DeleteAccountResponse.builder()
